+++ fractal_tasks_core/cellpose_segmentation.py
@@ -620,9 +620,10 @@ def cellpose_segmentation(
 if __name__ == "__main__":
 
     from pydantic import BaseModel
+    from pydantic import Extra
     from fractal_tasks_core._utils import run_fractal_task
 
-    class TaskArguments(BaseModel):
+    class TaskArguments(BaseModel, extra=Extra.forbid):
         # Fractal arguments
         input_paths: Sequence[str]
         output_path: str
@@ -636,7 +637,7 @@ if __name__ == "__main__":
         level: int
         relabeling: bool = True
         anisotropy: Optional[float] = None
-        diameter_level0: float
+        diameter_level0: Optional[float]
         cellprob_threshold: Optional[float]
         flow_threshold: Optional[float]
         ROI_table_name: Optional[str]