+++ fractal_tasks_core/create_ome_zarr.py
@@ -14,7 +14,6 @@ Copyright 2022 (C)
 Create structure for OME-NGFF zarr array
 """
 import os
-from glob import glob
 from pathlib import Path
 from typing import Any
 from typing import Dict
@@ -30,6 +29,7 @@ import fractal_tasks_core
 from fractal_tasks_core.lib_channels import check_well_channel_labels
 from fractal_tasks_core.lib_channels import define_omero_channels
 from fractal_tasks_core.lib_channels import validate_allowed_channel_input
+from fractal_tasks_core.lib_glob import glob_with_multiple_patterns
 from fractal_tasks_core.lib_metadata_parsing import parse_yokogawa_metadata
 from fractal_tasks_core.lib_parse_filename_metadata import parse_filename
 from fractal_tasks_core.lib_regions_of_interest import prepare_FOV_ROI_table
@@ -50,7 +50,7 @@ def create_ome_zarr(
     output_path: str,
     metadata: Dict[str, Any],
     image_extension: str = "tif",
-    image_glob_pattern: Optional[str] = None,
+    image_glob_patterns: Optional[list[str]] = None,
     allowed_channels: Sequence[Dict[str, Any]],
     num_levels: int = 2,
     coarsening_xy: int = 2,
@@ -73,8 +73,13 @@ def create_ome_zarr(
     :param input_paths: TBD (common to all tasks)
     :param output_path: TBD (common to all tasks)
     :param metadata: TBD (common to all tasks)
-    :param image_extension: Filename extension of images (e.g. `tif` or `png`)
-    :param image_glob_pattern: TBD
+    :param image_extension: Filename extension of images (e.g. `"tif"` or
+                            `"png"`)
+    :param image_glob_patterns: If specified, only parse images with filenames
+                                that match with all these patterns. Patterns
+                                must be defined as in
+                                https://docs.python.org/3/library/fnmatch.html,
+                                e.g. `image_glob_pattern=["*_B03_*"]`.
     :param num_levels: Number of resolution-pyramid levels
     :param coarsening_xy: Linear coarsening factor between subsequent levels
     :param allowed_channels: A list of channel dictionaries, where each channel
@@ -95,8 +100,6 @@ def create_ome_zarr(
         )
     if metadata_table.endswith(".csv") and not os.path.isfile(metadata_table):
         raise FileNotFoundError(f"Missing file: {metadata_table=}")
-    if image_glob_pattern:
-        raise NotImplementedError
 
     # Identify all plates and all channels, across all input folders
     plates = []
@@ -109,8 +112,15 @@ def create_ome_zarr(
 
     for in_path_str in input_paths:
         in_path = Path(in_path_str)
-        glob_expression = str(in_path) + f"/*.{image_extension}"
-        input_filenames = glob(glob_expression)
+
+        # Glob image filenames
+        patterns = [f"*.{image_extension}"]
+        if image_glob_patterns:
+            patterns.extend(image_glob_patterns)
+        input_filenames = glob_with_multiple_patterns(
+            folder=in_path_str,
+            patterns=patterns,
+        )
 
         tmp_wavelength_ids = []
         tmp_plates = []
@@ -133,16 +143,18 @@ def create_ome_zarr(
         tmp_wavelength_ids = sorted(list(set(tmp_wavelength_ids)))
 
         info = (
-            f"Listing all plates/channels from {glob_expression}\n"
+            "Listing plates/channels:\n"
+            f"Folder:   {in_path_str}\n"
+            f"Patterns: {patterns}\n"
             f"Plates:   {tmp_plates}\n"
             f"Channels: {tmp_wavelength_ids}\n"
         )
 
         # Check that only one plate is found
         if len(tmp_plates) > 1:
-            raise Exception(f"{info}ERROR: {len(tmp_plates)} plates detected")
+            raise ValueError(f"{info}ERROR: {len(tmp_plates)} plates detected")
         elif len(tmp_plates) == 0:
-            raise Exception(f"{info}ERROR: No plates detected")
+            raise ValueError(f"{info}ERROR: No plates detected")
         plate = tmp_plates[0]
 
         # If plate already exists in other folder, add suffix
@@ -208,13 +220,20 @@ def create_ome_zarr(
         if metadata_table == "mrf_mlf":
             mrf_path = f"{in_path}/MeasurementDetail.mrf"
             mlf_path = f"{in_path}/MeasurementData.mlf"
-            site_metadata, total_files = parse_yokogawa_metadata(
-                mrf_path, mlf_path
+
+            site_metadata, number_images_mlf = parse_yokogawa_metadata(
+                mrf_path,
+                mlf_path,
+                filename_patterns=image_glob_patterns,
             )
             site_metadata = remove_FOV_overlaps(site_metadata)
 
         # If a metadata table was passed, load it and use it directly
         elif metadata_table.endswith(".csv"):
+            logger.warning(
+                "Since a custom metadata table was provided, there will "
+                "be no additional check on the number of image files."
+            )
             site_metadata = pd.read_csv(metadata_table)
             site_metadata.set_index(["well_id", "FieldIndex"], inplace=True)
 
@@ -230,23 +249,43 @@ def create_ome_zarr(
         # Identify all wells
         plate_prefix = dict_plate_prefixes[plate]
 
-        plate_image_iter = glob(
-            f"{in_path}/{plate_prefix}_*.{image_extension}"
+        patterns = [f"{plate_prefix}_*.{image_extension}"]
+        if image_glob_patterns:
+            patterns.extend(image_glob_patterns)
+        plate_images = glob_with_multiple_patterns(
+            folder=str(in_path), patterns=patterns
         )
 
         wells = [
-            parse_filename(os.path.basename(fn))["well"]
-            for fn in plate_image_iter
+            parse_filename(os.path.basename(fn))["well"] for fn in plate_images
         ]
         wells = sorted(list(set(wells)))
 
         # Verify that all wells have all channels
         for well in wells:
-            well_image_iter = glob(
-                f"{in_path}/{plate_prefix}_{well}*.{image_extension}"
+            patterns = [f"{plate_prefix}_{well}_*.{image_extension}"]
+            if image_glob_patterns:
+                patterns.extend(image_glob_patterns)
+            well_images = glob_with_multiple_patterns(
+                folder=str(in_path), patterns=patterns
             )
+
+            # Check number of images matches with expected one
+            if metadata_table == "mrf_mlf":
+                num_images_glob = len(well_images)
+                num_images_expected = number_images_mlf[well]
+                if num_images_glob != num_images_expected:
+                    raise ValueError(
+                        f"Wrong number of images for {well=}\n"
+                        f"Expected {num_images_expected} (from mlf file)\n"
+                        f"Found {num_images_glob} files\n"
+                        "Other parameters:\n"
+                        f"  {image_extension=}\n"
+                        f"  {image_glob_patterns=}"
+                    )
+
             well_wavelength_ids = []
-            for fpath in well_image_iter:
+            for fpath in well_images:
                 try:
                     filename_metadata = parse_filename(os.path.basename(fpath))
                     well_wavelength_ids.append(
@@ -383,7 +422,7 @@ def create_ome_zarr(
         num_levels=num_levels,
         coarsening_xy=coarsening_xy,
         image_extension=image_extension,
-        image_glob_pattern=image_glob_pattern,
+        image_glob_patterns=image_glob_patterns,
         original_paths=input_paths[:],
     )
     return metadata_update
@@ -391,14 +430,15 @@ def create_ome_zarr(
 
 if __name__ == "__main__":
     from pydantic import BaseModel
+    from pydantic import Extra
     from fractal_tasks_core._utils import run_fractal_task
 
-    class TaskArguments(BaseModel):
+    class TaskArguments(BaseModel, extra=Extra.forbid):
         input_paths: Sequence[str]
         output_path: str
         metadata: Dict[str, Any]
         image_extension: str
-        image_glob_pattern: Optional[str]
+        image_glob_patterns: Optional[list[str]]
         allowed_channels: Sequence[Dict[str, Any]]
         num_levels: Optional[int]
         coarsening_xy: Optional[int]