+++ fractal_tasks_core/illumination_correction.py
@@ -104,7 +104,7 @@ def illumination_correction(
 ) -> Dict[str, Any]:
 
     """
-    FIXME
+    FIXME docstring
 
     Example inputs:
     input_paths: ["/some/path"]
@@ -273,9 +273,10 @@ def illumination_correction(
 if __name__ == "__main__":
 
     from pydantic import BaseModel
+    from pydantic import Extra
     from fractal_tasks_core._utils import run_fractal_task
 
-    class TaskArguments(BaseModel):
+    class TaskArguments(BaseModel, extra=Extra.forbid):
         # Fractal arguments
         input_paths: Sequence[str]
         output_path: str